refactor : setting profile을 profile setting으로 이름 변경
