+++ src/router/index.tsx
@@ -11,7 +11,7 @@ import Main from '@pages/Main';
 import NotFound from '@pages/NotFound';
 import Notice from '@pages/Notice';
 import Profile from '@pages/Profile';
-import SettingProfile from '@pages/SettingProfile';
+import ProfileSetting from '@pages/ProfileSetting';
 import ShareDetail from '@pages/ShareDetail';
 import ShareList from '@pages/ShareList';
 import ShareMap from '@pages/ShareMap';
@@ -36,7 +36,7 @@ export const routes = [
       { path: P.chatroomDetail + '/:id', element: <ChatroomDetail /> },
       { path: P.shareForm + '/:type', element: <ShareRegistration /> },
       { path: P.profile, element: <Profile /> },
-      { path: P.settingsProfile, element: <SettingProfile /> },
+      { path: P.profileSetting, element: <ProfileSetting /> },
       { path: P.editUserInfo, element: <EditUserInfo /> },
       { path: P.salesHistory, element: <History menuType='sales' /> },
       { path: P.purchaseHistory, element: <History menuType='purchase' /> },
